pyproject.toml
[project]
name = "byggesak-feed"
version = "0.1.0"
description = "Fetch municipal building cases from Asker kommune and generate Atom feeds"
authors = [
    { name = "Sven Ulland", email = "sven.ulland@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "requests",
    "pyyaml",
    "feedgen",
]

[project.scripts]
byggesak-feed = "byggesak_feed.cli:main"

[build-system]
requires = ["uv_build>=0.10.11,<0.11.0"]
build-backend = "uv_build"
